fix: rename parser
Renaming parser because it may overwrite predefined packages
